Change name of url parameter passed as argument to GET view.
The name of the parameter in the url mapped to the GET view
is now 'date__year'. This is the same query parameter that
we would use to perform a query with Django's ORM to retrieve
those entries from the database that correspond with the
specified year. This helps us leverage Django Rest Framework's
lookup functionality, since this query field will be automatically
used to form the intended queryset.