Finish initial implementation of CRUD API leveraging Rest Framework's mixins.
The implementation of the mixins is custozmied, since the operations that
we need to perform on create, update and delete go beyond the basic
functionalities that Rest Framework offers. Because of this changes,
it is of outmost importance that extensive tests are written later to
ensure that everything works as expected.