Adapt Day serializer to the needs of the app.
I've left uuid as the only read-only field of the model. The reason
is that leaving the user pk also as read-only makes it impossible
to later perform database transactions that need this information,
since Rest Framework ignores read-only fields.
I've removed the owner field since I don't think we need it in this
application.